auth: pemisahan controller auth
