+++ app/Http/Controllers/Auth/AuthController.php
@@ -12,73 +12,6 @@ use Illuminate\Support\Facades\Hash;
 class AuthController extends Controller
 {
     /**
-     * Menampilkan form login.
-     */
-    public function showLoginForm()
-    {
-        return view('auth.login');
-    }
-
-    /**
-     * Melakukan proses login.
-     */
-    public function login(Request $request)
-    {
-        // Validasi input
-        $credentials = $request->validate([
-            'email' => 'required|email',
-            'password' => 'required|string|min:8',
-        ]);
-
-        // Proses login
-        if (Auth::attempt($credentials)) {
-            $request->session()->regenerate();
-            Cache::put('success', 'Login berhasil! Selamat datang!', now()->addSeconds(5));
-            return redirect()->intended('/');
-        }
-        return back()->withErrors([
-            'email' => 'Email atau password salah.',
-        ])->withInput();
-    }
-
-    /**
-     * Menampilkan form register.
-     */
-    public function showRegisterForm()
-    {
-        return view('auth.register');
-    }
-
-    /**
-     * Melakukan proses registrasi pengguna baru.
-     */
-    public function register(Request $request)
-    {
-        // Validasi input registrasi
-        $request->validate([
-            'name' => 'required|string|max:255',
-            'email' => 'required|string|email|max:255|unique:users',
-            'password' => 'required|string|min:8|confirmed',
-            'no_telepon' => 'required|string|max:15|unique:users', // Menambahkan validasi untuk nomor telepon
-        ]);
-
-        // Membuat pengguna baru
-        $user = User::create([
-            'name' => $request->name,
-            'email' => $request->email,
-            'password' => Hash::make($request->password),
-            'no_telepon' => $request->no_telepon, // Menyimpan nomor telepon
-            'role' => 'user',
-        ]);
-
-        // Login otomatis setelah registrasi
-        Auth::login($user);
-
-        Cache::put('success', 'Registrasi berhasil! Selamat datang!', now()->addSeconds(5));
-        return redirect('/');
-    }
-
-    /**
      * Melakukan logout.
      */
     public function logout(Request $request)
